recipes/fdb/5.18/meta/private/pyproject.toml
[tool.poetry]
name = "nwp-demo"
# Keep only Major and Minor version in this configuration,
# the patch version is calculated when releasing the library
version = "1.0"
description = "Demo scripts and notebooks for reading and processing forecast data using Polytope."
authors = ["MeteoSwiss"]
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
meteodata-lab = {version = "0.6.0", extras = ["polytope", "regrid"]}
# version > 0.1.1 of pyfdb requires eccodes >= 2.39
pyfdb = "0.1.2"
fdb-utils = { git = "https://github.com/MeteoSwiss/fdb-utils.git", tag = "v1.1.5"}
ipykernel = "*"
earthkit-plots = "*"
earthkit-geo = "*"
earthkit-meteo = "*"
earthkit-data = ">=0.16.3"
pygribjump = { git = "https://github.com/ecmwf/gribjump.git", tag = "0.10.2" }
polytope-client = "^0.7.8"
nbconvert = "*"
pyyaml = "*"
kaleido = "^1.1.0"

[tool.pytest.ini_options]
testpaths = ["test"]
pythonpath = "."

[tool.yapf]
based_on_style = "pep8"
column_limit = "120"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
